requirements.txt
Flask==2.1.2
flask-marshmallow
Flask-SQLAlchemy
Alembic
marshmallow-sqlalchemy
pytest